Refactoring _timeoutWrapper into command.py
