+++ IOT Integration Code/MicroPython/Integration Code/Code/Commands/Commands.py
@@ -9,20 +9,9 @@ from .command import Command as CommandClass
 prebuilt = dict()
 
 
-async def _timeoutWrapper(coroutine):
-    async def __timeoutWrapper(*_args, time=..., **_kwargs):
-        if time is not ...:
-            await asio.wait_for(coroutine(*_args, **_kwargs), timeout=time)
-        else:
-            print(
-                f"#>\tRunning __timeoutWrapper with no given time, defaulting to asio.gather(...)")
-            await asio.gather(coroutine(*_args, **_kwargs))
-    return __timeoutWrapper
-
-
 def importSet(commands):
     for name, command in commands.items():
-        prebuilt[name] = _timeoutWrapper(command)
+        prebuilt[name] = command
         if command.name is None:
             command.name = name  # Set name of command to handle used to call it :)
 
@@ -43,13 +32,13 @@ async def _execute(commands, time=..., **overflow):
     for command in commands:
         if type(command) is not CommandClass:
             print("## ERROR: Command is not of type CommandClass")
-            continue
+            return
         print(
             f"#\tMaking Task {str(command.name)+' ' if hasattr(command,'name') else '__'}...")
     if time is ...:
         print(f"#> Note: {time=}, is not set, running until completion")
     print(f"## Holding tasks ({time} seconds) ...")
-    await asio.gather(*[task.asioDo(time=time) for task in commands])
+    await asio.gather(*[command.asioDo(time=time) for command in commands])
     print("## Released tasks.")
